on_call_medical/requirements.txt
asgiref==3.3.1
Django==3.1.5
django-cors-headers==3.6.0
django-phonenumber-field==5.0.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
phonenumbers==8.12.16
PyJWT==2.0.1
pytz==2020.5
sqlparse==0.4.1
